#2373: Improve global loader (#2372)
