+++ components/pages/submission-system/create-program/index.tsx
@@ -38,7 +38,7 @@ import CREATE_PROGRAM_MUTATION from './CREATE_PROGRAM_MUTATION.gql';
 import { useMutation } from '@apollo/react-hooks';
 import useCommonToasters from 'components/useCommonToasters';
 import SIDE_MENU_PROGRAM_LIST from '../SIDE_MENU_PROGRAM_LIST.gql';
-import { useGlobalLoadingState } from 'components/ApplicationRoot';
+import useGlobalLoader from 'components/GlobalLoader';
 
 /* *************************************** *
  * Reshape form data for gql input
@@ -87,11 +87,11 @@ const CreateProgramPage = () => {
     },
   });
 
-  const { setLoading: setFormDisabled } = useGlobalLoadingState();
+  const { setGlobalLoading } = useGlobalLoader();
 
   const onSubmit = async (data) => {
     try {
-      setFormDisabled(true);
+      setGlobalLoading(true);
       await sendCreateProgram({
         variables: { program: createProgramInput(data) },
       });
@@ -118,7 +118,7 @@ const CreateProgramPage = () => {
         commonToasters.unknownError();
       }
     } finally {
-      setFormDisabled(false);
+      setGlobalLoading(false);
     }
   };
