+++ components/pages/submission-system/program-clinical-submission/PageContent.tsx
@@ -43,7 +43,8 @@ import { displayDateAndTime } from 'global/utils/common';
 import { capitalize } from 'global/utils/stringUtils';
 import { useToaster } from 'global/hooks/toaster';
 import ErrorNotification, { getDefaultColumns } from '../ErrorNotification';
-import { ModalPortal, useGlobalLoadingState } from 'components/ApplicationRoot';
+import useGlobalLoader from 'components/GlobalLoader';
+import ModalPortal from 'components/Modal';
 import SignOffValidationModal from './SignOffValidationModal';
 import SubmissionSummaryTable from './SubmissionSummaryTable';
 import useUserConfirmationModalState from './useUserConfirmationModalState';
@@ -112,7 +113,7 @@ const gqlClinicalEntityToClinicalSubmissionEntityFile =
 
 const PageContent = () => {
   const { shortName: programShortName } = usePageQuery<{ shortName: string }>();
-  const { setLoading: setPageLoaderShown } = useGlobalLoadingState();
+  const { setGlobalLoading } = useGlobalLoader();
 
   // not declared as a side effect that changes with program change
   // change in 'data' always seems to take precedence over currentFileList within `defaultClinicalEntityType`
@@ -372,7 +373,7 @@ const PageContent = () => {
     try {
       const userDidApprove = await getSignOffConfirmation();
       if (userDidApprove) {
-        setPageLoaderShown(true);
+        setGlobalLoading(true);
         await sleep();
         const { data: newData } = await signOffSubmission({
           variables: {
@@ -394,13 +395,13 @@ const PageContent = () => {
               'Your clinical data has been submitted. You will see the updates on your dashboard shortly.',
           });
         } else {
-          setPageLoaderShown(false);
+          setGlobalLoading(false);
         }
       }
     } catch (err) {
       await refetch();
       commonToaster.unknownErrorWithReloadMessage();
-      setPageLoaderShown(false);
+      setGlobalLoading(false);
     }
   };
