+++ components/pages/submission-system/program-sample-registration/Instructions/RegisterSamplesModal.tsx
@@ -19,7 +19,8 @@
 
 import React from 'react';
 import Modal from '@icgc-argo/uikit/Modal';
-import { ModalPortal, useGlobalLoadingState } from 'components/ApplicationRoot';
+import useGlobalLoader from 'components/GlobalLoader';
+import ModalPortal from 'components/Modal';
 import { useMutation } from '@apollo/react-hooks';
 import pluralize from 'pluralize';
 import COMMIT_CLINICAL_REGISTRATION_MUTATION from './COMMIT_CLINICAL_REGISTRATION_MUTATION.gql';
@@ -62,14 +63,14 @@ export default function RegisterSamplesModal({
     ],
   });
 
-  const { setLoading: setGlobalLoadingState } = useGlobalLoadingState();
+  const { setGlobalLoading } = useGlobalLoader();
 
   const toaster = useToaster();
 
   const handleActionClick = async () => {
     handleCancelClick();
 
-    setGlobalLoadingState(true);
+    setGlobalLoading(true);
     await sleep();
 
     await commitRegistration()
@@ -105,7 +106,7 @@ export default function RegisterSamplesModal({
           content: error.toString(),
         });
       });
-    setGlobalLoadingState(false);
+    setGlobalLoading(false);
   };
 
   return (
