+++ pages/file/[fileId].tsx
@@ -26,7 +26,7 @@ import sqonBuilder from 'sqon-builder';
 import { useQuery } from '@apollo/react-hooks';
 import VALID_FILE_ENTITY_CHECK from './VALID_FILE_ENTITY_CHECK.gql';
 import get from 'lodash/get';
-import { useGlobalLoadingState } from 'components/ApplicationRoot';
+import useGlobalLoader from 'components/GlobalLoader';
 
 export default createPage({
   isPublic: true,
@@ -54,18 +54,18 @@ export default createPage({
 
   const isValidEntity = !!get(data, 'file.hits.total', false);
 
-  const { setLoading: setLoaderShown, isLoading: isLoaderShown } = useGlobalLoadingState();
+  const { setGlobalLoading } = useGlobalLoader();
 
   if (!loading && !isValidEntity) {
-    setLoaderShown(false);
+    setGlobalLoading(false);
     const err = new Error('Page Not Found') as Error & { statusCode?: number };
     err[ERROR_STATUS_KEY] = 404;
     return <ErrorPage statusCode={404} />;
   } else if (loading) {
-    setLoaderShown(true);
+    setGlobalLoading(true);
     return <div></div>;
   }
-  setLoaderShown(false);
+  setGlobalLoading(false);
 
   return <FileEntityPage {...props} fileId={fileId} />;
 });
